Improve substepping handling transvisotrop mat
- Removed try-catch handling of the substepping within
InelasticDefgradTransvIsotropElastViscoplast
- Improved substepping management by using an error status
integer indicating which of the possible errors is encountered.
Thus, substepping is now triggered using this error status. The
same goes for the perturbation-based linearization, which is
triggered if an error is encountered in the analytical linearization.
- Also optimized the different management of log substepping and
standard substepping within the viscoplasticity laws, by using a
check_overflow variable instead of setting an extremely small
number for check_dt.
- Also added overflow error check for the derivative evaluation in
the implemented Reformulated Johnson-Cook law, by first calculating the
logarithm of the derivatives. The error checking is now also done for
logarithmic substepping, but with other numerical bounds.