Add analysis utils trv.iso. material
- Added analysis utilities for the time integration within
InelasticDefgradTransvIsotropElastViscoplast. Specifically, a dedicated
object is used to track various key performance factors, such as the
number of iterations, the number of substeps, or the various error types
arising during the Local Newton Loop.
- A csv writer is integrated within the TimIntAnalysisUtils object,
which writes runtime output for the analysis of the time integration
scheme.
- The user can select whether to write dedicated runtime output for the
time integration via a dedicated boolean variable.
- Added a new dedicated error management functions to avoid repeating
code.
- Adapted result description for the framework test of the Anand law.