Adapt time integr. features for trv.iso. material
- Improved performance by adding partial evaluation types for state
quantities and their derivative. This is especially important in the
cases where only the plastic strain rate and its derivatives must be
computed, and the remaining state variables (e.g. stretching tensor) are
not realy relevant.
- Fixed bug regarding the second state evaluation in the predictor
adaptation. Also, the predictor adaptation now also accounts for the
derivatives of state quantities, which can also lead to overflow errors
and might require readjusting the predictor.
- Added the maximum plastic strain increment and the derivative
increment as input parameters.
- Improved the error management in the predictor adaptation routine, by
also accounting for overflow errors that may appear while integrating
the plastic strain.