Fix analytical linearization
- Fixed bug in analytical linearization, where the current plasticity
state was not evaluated and led to 0 contribution to the stiffness
tensor.
- Added debugging output for the 1D tensor interpolation. THIS SHOULD BE
REMOVED in the future.
- Improved the debugging framework by adding vectors of ele_gid and gp
for which we want output. Also added further details for debugging
output.
- Changed the maximum plastic strain increment and the derivative
increment in the unit tests. THIS SHOULD BE CHANGED BACK to the original
when I plan to merge the branch into 4C.