Fix R determination in tensor interpolation
- Fixed the determination of the relative R rotations in the tensor
interpolation routine. In the previous structure, the base matrix was
not necessarily decomposed at the time where the relative rotations were
determined, which led to the base matrix being 0 instead of its correct
value.
- Fixed bug in the evaluation of the exponential weighting function of
the tensor interpolation scheme (squared distance!).
- Added another unit test specifically for the 1D interpolation, which
more thoroughly checks the inner workings of the tensor interpolation.
- Fixed unit tests of the matrix logarithm evaluations.