Enhance tensor interpolation with interp-types
- Added various types for rotation and eigenvalue
interpolation, although we currently only concentrate on R-LOG. The
other ones which are added but not implemented can be found in Satheesh
et al. 2024.
- Adapted unit tests of tensor interpolation.
- The tensor interpolator within
InelasticDefgradTransvIsotropElastViscoplast is now initialized via a
dedicated function. Currently, the parameters of the tensor interpolator
are set.
- Some changes were made to the dummy tests for the
transversely-isotropic material - also, some parameters have changed,
such that the specific unit tests may fail.