Fix redundant preevaluation
- The pre_evaluate method is called within the redundant material
evaluation as described in Issue #121. This means, that history
variables and other "complicated" variables are reset within this
routine, which takes place before ::update.
This is undesired behavior for some of the variables, such as the
predictor interpolation factors - it resets the calculated current_xi
to 0 (when boost was not used).
- This was fixed by adding a dedicated method prepare_non_repeat_tasks,
designed to perform the non-repeatable preevaluation tasks directly
within evaluate_inverse_inelastic_defgrad, and after checking whether we
have a redundant evaluation call. This fixed the issue above.
- Added GP output for the predictor interpolation factors xi and max_xi.