Add csv output routine for failed LNL
- Added a csv output routine if a LNL fails, containing all
relevant data as in the Gauss point output, but for a single
GP of an element this time (failing LNL).
- Moved iteration status and its conversions to the utilities
file.
- In evaluate_state_quantities: we now return the evaluated
state quantities in case of errors, as opposed to returning
empty objects. This is better for the LNL iteration data,
otherwise we see only 0 stresses if an iteration fails.