Replace comput. of plastic strain in pred-adapt
- Replaced the computation of the plastic strain for the adapted
predictor. Specifically, we now integrate the plastic strain via
the flow rule only for the plastic predictor, and interpolate
the plastic strain between the predictor extrema (scalar
interpolation) in the same manner as we do the tensor interpolation
for the plastic deformation gradients. After Abhi reworks all
interpolation schemes, we have to rectify the hard-coded parts.
Maybe it even makes sense to not have a stored tensor interpolator
object, but only initialize it in the predictor adaptation routine
analogous to how we currently handle the scalar interpolation.
- All unit tests are currently passed. It remains to be determined
whether we reach the same speed-ups in our simulations as with the
previous "inconsistent" approach of always integrating the flow rule.