Remove 2nd boost evaluation of elastic predictor
- Removed the 2nd elastic predictor evaluation, which always
evaluated the elastic predictor in the second microiteration of
the predictor adaptation when using performance boosting
strategies. However, we shall obtain (better) or same results,
since xi_l = 0 is evaluated in the second iteration if the
boost initial point xi_boost_init leads to an under-yield-surface
state.
- Improved documentation in the routines for the LNL, the predictor
adaptation and the line search.
- Added notes for the documentation of the residual and jacobian
computations, to specify that state quantities and their derivatives
are modified when calling the function.