Fix strategy for checking elast. pred
- Fixed the strategy for checking the
elastic predictor in the predictor adaptation:
either when the user specifically requires it,
or when all interpolation factors are 0.0 anyway.
Otherwise we would evaluate this while also updating
the plastic strain, which in some cases leads
to the "under_the_yield_surface" error, which
completely halts the predictor adaptation.
- Renamed and relinked things to the general viscoplastic
law in the dummy unit test used for debugging, to
not explicitly use ref. JC.