Finalised improved interpolation
- We introduce different stretch and rotation assignment
types for the plastic predictor. Depending on the rotation
specification, either the inverse plastic or the elastic
deformation gradient are decomposed and interpolated.
- Improved interpolation framework, now called
LocalNewtonGuessInterpolation:
	* Dedicated PredictorDecomposition and
	  InterpolationPoint objects used to interpolate
	  in the formed interpolation space.
	* Improved check: is interpolation still possible?
	  Now we check the minimum upper - lower interval,
 	  the number of repredictorization, and the number
	  of current interpolation iterations.
	* While the eigenvector rotation is still interpolated
	  linearly, we make the eigenvalue interpolation
	  consistent in regards to determinant preservation,
	  by using the logarithmic weighted average
	  method from Abhi's paper. Linear weighting is used
	  between the interpolation points related to both
	  predictors. We still only interpolate two of the
	  eigenvalues, and the third is obtained from the
	  determinant (which is either 1 or detF, depending
	  on the interpolated deformation gradient type).
- Implemented dedicated function to verify whether a combination
inverse plastic defgrad + plastic strain is a valid LNL initial
guess. It should be numerically evaluable (both residual and
Jacobian), and should lead to plastic strain.
- Improved reinterpolation strategy: we now "go back" towards the
elastic predictor first. If the new interpolation point is too near
to the lower bound (user-specified tolerance), we set lower <- new,
and try to determine an initial guess in the herewith updated
interpolation interval. The same goes for the case where the new
interpolation point is not a viable initial guess. If it is however
a viable initial guess, it is directly accepted and the Local Newton
on.
- Our investigations show consistent R-rotation behavior
(solution = elastic predictor = plastic predictor). Interestingly,
the Q-rotation of the solution is always the same as the one from
the elastic predictor. Further investigations? Can we improve the
interpolation algorithm further?